Move website into a subdirectory Replace jest plugin with npm version
