--- .netlify/netlify-plugin-jest/manifest.yml
@@ -1,2 +1,0 @@
-name: netlify-plugin-jest
-inputs: []
