+++ package.json
@@ -7,17 +7,7 @@
     "url": "https://github.com/seancdavis/ssg-build-performance-tests"
   },
   "scripts": {
-    "build": "npm-run-all clean build:css build:html build:js",
-    "build:css": "postcss src/assets/css/main.css -o dist/assets/css/bundle.css",
-    "build:html": "eleventy",
-    "build:js": "webpack",
-    "clean": "rm -rf dist",
-    "develop": "npm-run-all --parallel develop:html develop:js develop:css",
-    "develop:css": "postcss src/assets/css/main.css -o dist/assets/css/bundle.css --watch",
-    "develop:html": "eleventy --serve --port 8000",
-    "develop:js": "webpack",
     "heroku-postbuild": "echo \"HEROKU: NO BUILD NECESSARY.\"",
-    "import:results": "node lib/import",
     "start": "node index.js",
     "test:builds": "node lib/run",
     "test": "jest -i"
@@ -26,13 +16,9 @@
   "author": "Sean C Davis",
   "license": "MIT",
   "dependencies": {
-    "@11ty/eleventy": "^0.11.0",
     "@apollo/client": "^3.2.5",
     "apollo-boost": "^0.4.9",
-    "autoprefixer": "^9.8.6",
     "chalk": "^4.1.0",
-    "chart.js": "^2.9.3",
-    "cssnano": "^4.1.10",
     "express": "^4.17.1",
     "glob": "^7.1.6",
     "graphql": "^15.4.0",
@@ -41,17 +27,9 @@
     "lodash": "^4.17.19",
     "lorem-ipsum": "^2.0.3",
     "node-fetch": "^2.6.1",
-    "npm-run-all": "^4.1.5",
     "object-path": "^0.11.8",
-    "postcss-cli": "^7.1.1",
-    "postcss-custom-properties": "^10.0.0",
-    "postcss-import": "^12.0.1",
-    "postcss-nested": "^4.2.3",
     "progress": "^2.0.3",
     "rimraf": "^3.0.2",
-    "tailwindcss": "^1.7.3",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
     "yargs": "^16.0.3"
   }
 }
