+++ website/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "website",
+  "version": "0.0.1",
+  "description": "Data visualization for SSG comparison.",
+  "scripts": {
+    "build": "npm-run-all clean build:css build:html build:js",
+    "build:css": "postcss src/assets/css/main.css -o dist/assets/css/bundle.css",
+    "build:html": "eleventy",
+    "build:js": "webpack",
+    "clean": "rm -rf dist",
+    "develop": "npm-run-all --parallel develop:html develop:js develop:css",
+    "develop:css": "postcss src/assets/css/main.css -o dist/assets/css/bundle.css --watch",
+    "develop:html": "eleventy --serve --port 8000",
+    "develop:js": "webpack",
+    "import:results": "node lib/import"
+  },
+  "keywords": [],
+  "author": "Sean C Davis",
+  "license": "MIT",
+  "dependencies": {
+    "@11ty/eleventy": "^0.11.0",
+    "@seancdavis/netlify-plugin-jest": "^0.1.0",
+    "autoprefixer": "^9.8.6",
+    "chart.js": "^2.9.3",
+    "cssnano": "^4.1.10",
+    "graphql-tag": "^2.11.0",
+    "lodash": "^4.17.19",
+    "npm-run-all": "^4.1.5",
+    "postcss-cli": "^7.1.1",
+    "postcss-custom-properties": "^10.0.0",
+    "postcss-import": "^12.0.1",
+    "postcss-nested": "^4.2.3",
+    "tailwindcss": "^1.7.3",
+    "webpack": "^4.44.1",
+    "webpack-cli": "^3.3.12"
+  }
+}
